Table of Songs added
The songs and their attributes will be displayed in a table of a fixed size.

Bye using '%n', and the printf() method, i was able to clearly show contents of the song arraylist.

This looks cleaner than previously where they were disorganised and had no borders or columns to separate them.