Song objects updated with new attribute and can be used.
Each song object will have a URL link, in the form of a string, associated with it. This gives it a unique location for the song to be found online.

Between lines 219 and 249, the remove option has been used as a template in order to search and find a song according to the title and artist. This gives the option of playing the specific song found.

By default, the program will ask if the user would want to play another in song in the case that it does not work the first time or they return to the program from the link.